+++ pom.xml
@@ -45,7 +45,7 @@
     <hadoop.version>3.1.0</hadoop.version>
     <hive.execution.engine>mr</hive.execution.engine>
     <hive.version>3.1.2</hive.version>
-    <junit.jupiter.version>5.6.0</junit.jupiter.version>
+    <junit.jupiter.version>5.7.0</junit.jupiter.version>
     <license.maven.plugin.version>3.0</license.maven.plugin.version>
     <tez.version>0.9.1</tez.version>
   </properties>
@@ -159,6 +159,13 @@
       <artifactId>junit-jupiter</artifactId>
       <version>${junit.jupiter.version}</version>
     </dependency>
+
+    <dependency>
+      <groupId>org.junit.vintage</groupId>
+      <artifactId>junit-vintage-engine</artifactId>
+      <version>${junit.jupiter.version}</version>
+    </dependency>
+
   </dependencies>
 
   <build>
@@ -167,7 +174,7 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-surefire-plugin</artifactId>
-        <version>2.21.0</version>
+        <version>2.22.2</version>
         <configuration>
           <!-- Hiverunner need this for some queries (property -XX:MaxPermSize
             was removed in java 8) -->
