+++ src/test/java/com/klarna/hiverunner/examples/HelloAnnotatedHiveRunnerTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2013-2020 Klarna AB
+ * Copyright (C) 2013-2021 Klarna AB
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,10 +26,12 @@ import com.klarna.hiverunner.annotations.HiveSetupScript;
 import com.klarna.hiverunner.config.HiveRunnerConfig;
 import org.apache.commons.collections.MapUtils;
 import org.junit.Assert;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.io.File;
+import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -41,7 +43,7 @@ import java.util.Map;
  * All HiveRunner tests should run with the StandaloneHiveRunner
  */
 @ExtendWith(HiveRunnerExtension.class)
-public class HelloAnnotatedHiveRunner {
+public class HelloAnnotatedHiveRunnerTest {
 
     /**
      * Explicit test class configuration of the HiveRunner runtime.
@@ -100,6 +102,13 @@ public class HelloAnnotatedHiveRunner {
             "helloHiveRunner/create_ctas.sql"
     }, encoding = "UTF-8")
     private HiveShell hiveShell;
+
+    
+    @BeforeEach
+    public void setupSourceDatabase() {
+        hiveShell.execute(Paths.get("src/test/resources/helloHiveRunner/create_table.sql"));
+        hiveShell.execute(Paths.get("src/test/resources/helloHiveRunner/create_ctas.sql"));
+    }
 
     @Test
     public void testTablesCreated() {
